minor refactor, clean unused
